+++ client/src/App.tsx
@@ -1,181 +1,32 @@
+// App.tsx
 import './App.css';
-import { CreateListingPage } from './pages/CreateListingPage';
+import { Route, Routes } from 'react-router-dom';
+import Nav from './components/Nav';
+import { Footer } from './components/Footer';
 import { Home } from './pages/Home';
 import { LoginPage } from './pages/LoginPage';
-import { Route, Routes } from 'react-router-dom';
+import { SignUp } from './pages/SignUp';
+import { CreateListingPage } from './pages/CreateListingPage';
 import { Shop } from './pages/Shop';
 import { ProductDetailsPage } from './pages/ProductDetailsPage';
 import { ShoppingCartPage } from './pages/ShoppingCartPage';
-import { useEffect, useState } from 'react';
-import { AppContext } from './components/AppContext';
-import { CartItemsProps, Product, User } from './types/types';
 import { SellerDashboard } from './pages/SellerDashboard';
 import { ListingDetailsPage } from './pages/ListingDetailsPage';
 import { CheckoutPage } from './pages/CheckoutPage';
 import { OrderConfirmationPage } from './pages/OrderConfirmationPage';
 import { GenreCatalogPage } from './pages/GenreCatalogPage';
-import Nav from './components/Nav';
-import { Footer } from './components/Footer';
-import { SignUp } from './pages/SignUp';
+import { AppProvider } from './components/AppProvider';
 
 export default function App() {
-  const [cartItems, setCartItems] = useState<CartItemsProps[]>([]);
-  const [token, setToken] = useState<string>();
-  const [user, setUser] = useState<User>();
-
-  useEffect(() => {
-    async function loadCart() {
-      try {
-        const getToken = localStorage.getItem('token');
-        if (!getToken) {
-          return;
-        }
-        const res = await fetch(`/api/cart`, {
-          headers: {
-            Authorization: `Bearer ${getToken}`,
-          },
-        });
-        if (!res.ok) throw new Error(`Error: ${res.status}`);
-        const result = await res.json();
-        setCartItems(result);
-      } catch (error) {
-        console.error(error);
-      }
-    }
-    loadCart();
-  }, [token]);
-
-  useEffect(() => {
-    const token = localStorage.getItem('token');
-    const user = localStorage.getItem('user');
-    if (token && user) {
-      signIn(JSON.parse(user), token);
-    }
-  }, []);
-
-  async function addToCart(product: Product) {
-    try {
-      if (!token) {
-        alert('Sign in or make an account to purchase records!');
-      }
-      const response = await fetch('/api/cart/add', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-        body: JSON.stringify({
-          recordId: product?.recordId,
-        }),
-      });
-
-      if (!response.ok) throw new Error(`Error: ${response.status}`);
-      const result = await response.json();
-      setCartItems([...cartItems, result]);
-    } catch (error) {
-      console.error('Error adding to cart:', error);
-    }
-  }
-
-  async function signIn(user: User, token: string) {
-    localStorage.setItem('token', token);
-    localStorage.setItem('user', JSON.stringify(user));
-    setUser(user);
-    setToken(token);
-  }
-  async function signOut() {
-    localStorage.removeItem('token');
-    localStorage.removeItem('user');
-    setUser(undefined);
-    setToken(undefined);
-  }
-
-  async function removeFromCart(itemId: number) {
-    try {
-      const response = await fetch(`/api/cart/remove/${itemId}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      const result = await response.json();
-      const newCart = cartItems.filter(
-        (item) => item.itemsId !== result.itemsId
-      );
-      setCartItems(newCart);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-
-  async function deleteListing(recordId: number) {
-    const response = await fetch(`/api/delete-listing/${recordId}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
-    if (!response.ok) {
-      if (response.status === 421) {
-        throw new Error('421');
-      } else {
-        throw new Error('An error occurred');
-      }
-    }
-  }
-
-  async function handleCheckout() {
-    try {
-      const cartResponse = await fetch(`/api/cart/all/${user?.userId}`, {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      await cartResponse.json();
-
-      for (const cartItem of cartItems) {
-        const recordResponse = await fetch(
-          `/api/delete-record/${cartItem.recordId}`,
-          {
-            method: 'DELETE',
-            headers: {
-              'Content-Type': 'application/json',
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-        if (!recordResponse.ok)
-          console.error('An error occurred', recordResponse.status);
-      }
-      setCartItems([]);
-    } catch (error) {
-      console.error(error);
-    }
-  }
-  const contextValue = {
-    cartItems,
-    addToCart,
-    removeFromCart,
-    signIn,
-    signOut,
-    user,
-    token,
-    deleteListing,
-    handleCheckout,
-  };
-
   return (
-    <div>
-      <AppContext.Provider value={contextValue}>
+    <AppProvider>
+      <div>
         <Nav />
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="login" element={<LoginPage />} />
           <Route path="sign-up" element={<SignUp />} />
-          <Route path="CreateListing" element={<CreateListingPage />} />
+          <Route path="create" element={<CreateListingPage />} />
           <Route path="shop" element={<Shop />} />
           <Route
             path="ProductDetailsPage/:recordId"
@@ -195,7 +46,7 @@ export default function App() {
           <Route path="genre/:genreId" element={<GenreCatalogPage />} />
         </Routes>
         <Footer />
-      </AppContext.Provider>
-    </div>
+      </div>
+    </AppProvider>
   );
 }
