+++ client/src/pages/SellerDashboard.tsx
@@ -1,10 +1,8 @@
 import { ActiveListings } from '../components/ActiveListings';
-import { DashboardHeader } from '../components/DashboardHeader';
 
 export function SellerDashboard() {
   return (
     <>
-      <DashboardHeader />
       <ActiveListings />
     </>
   );
